All analyzer warnings fixed

+++ lib/src/core/Color.dart
@@ -13,9 +13,9 @@ class Color {
   num g;
   num b;
 
-  int get _rr => (r*255).floor().toInt();
-  int get _gg => (g*255).floor().toInt();
-  int get _bb => (b*255).floor().toInt();
+  int get _rr => (r*255).floor();
+  int get _gg => (g*255).floor();
+  int get _bb => (b*255).floor();
 
   Color( [num hex] )
       : r = 1.0,
